requirements.txt
jinja2-cli
pyyaml
docopt
paramiko
websocket-client
ansible
requests
